requirements.txt
alembic==1.0.0
blinker==1.4
Bootstrap-Flask==1.0.8
Click==7.0
dominate==2.3.4
Faker==0.9.1
Flask==1.0.2
Flask-Bootstrap==3.3.7.1
Flask-CKEditor==0.4.2
Flask-Dropzone==1.5.3
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.2.1
Flask-Moment==0.6.0
Flask-Redis==0.3.0
Flask-SQLAlchemy==2.3.2
flask-whooshee==0.6.0
Flask-WTF==0.14.2
gunicorn==19.9.0
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
Pillow==5.3.0
python-dateutil==2.7.3
python-dotenv==0.9.1
python-editor==1.0.3
python-slugify==1.2.6
redis==2.10.6
six==1.11.0
SQLAlchemy==1.2.12
text-unidecode==1.2
Unidecode==1.0.22
visitor==0.1.3
Werkzeug==0.14.1
Whoosh==2.7.4
WTForms==2.2.1
